@@ -4,6 +4,7 @@ namespace Oro\Bundle\UserBundle\Command;
 
 use Doctrine\DBAL\DBALException;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\NoResultException;
 
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputInterface;
@@ -35,12 +36,19 @@ class CreateUserCommand extends ContainerAwareCommand
             ->setName('oro:user:create')
             ->setDescription('Create user.')
             ->addOption('user-role', null, InputOption::VALUE_REQUIRED, 'User role')
-            ->addOption('user-business-unit', null, InputOption::VALUE_REQUIRED, 'User business unit')
-            ->addOption('user-name', null, InputOption::VALUE_REQUIRED, 'User name')
-            ->addOption('user-email', null, InputOption::VALUE_REQUIRED, 'User email')
+            ->addOption('user-business-unit', null, InputOption::VALUE_REQUIRED, 'User business unit (required)')
+            ->addOption('user-name', null, InputOption::VALUE_REQUIRED, 'User name (required)')
+            ->addOption('user-email', null, InputOption::VALUE_REQUIRED, 'User email (required)')
             ->addOption('user-firstname', null, InputOption::VALUE_REQUIRED, 'User first name')
             ->addOption('user-lastname', null, InputOption::VALUE_REQUIRED, 'User last name')
-            ->addOption('user-password', null, InputOption::VALUE_REQUIRED, 'User password');
+            ->addOption('user-password', null, InputOption::VALUE_REQUIRED, 'User password (required)')
+            ->addOption(
+                'user-organizations',
+                null,
+                InputOption::VALUE_IS_ARRAY | InputOption::VALUE_REQUIRED,
+                'User organizations'
+            )
+        ;
     }
 
     /**
@@ -48,18 +56,44 @@ class CreateUserCommand extends ContainerAwareCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        /** @var User $user */
         $user = $this->getUserManager()->createUser();
         $user->setEnabled(true);
 
         $options = $input->getOptions();
 
         try {
-            $this->updateUser($user, $options);
+            $this
+                ->checkRequiredOptions($options)
+                ->updateUser($user, $options);
         } catch (InvalidArgumentException $exception) {
             $output->writeln($exception->getMessage());
         } catch (DBALException $exception) {
             $output->writeln('User exists');
         }
+    }
+
+    /**
+     * @param array $options
+     * @throws InvalidArgumentException
+     * @return $this
+     */
+    protected function checkRequiredOptions($options)
+    {
+        $requiredOptions = [
+            'user-business-unit',
+            'user-name',
+            'user-email',
+            'user-password'
+        ];
+
+        foreach ($requiredOptions as $requiredOption) {
+            if (empty($options[$requiredOption])) {
+                throw new InvalidArgumentException('--' . $requiredOption . ' option required');
+            }
+        }
+
+        return $this;
     }
 
     /**
@@ -69,32 +103,6 @@ class CreateUserCommand extends ContainerAwareCommand
      */
     protected function updateUser(User $user, $options)
     {
-        if (!empty($options['user-role'])) {
-            $role = $this->getEntityManager()
-                ->getRepository('OroUserBundle:Role')
-                ->findOneBy(array('role' => $options['user-role']));
-
-            if (!$role) {
-                throw new InvalidArgumentException('Invalid Role');
-            }
-
-            $user->addRole($role);
-        }
-
-        if (!empty($options['user-business-unit'])) {
-            $businessUnit = $this->getEntityManager()
-                ->getRepository('OroOrganizationBundle:BusinessUnit')
-                ->findOneBy(array('name' => $options['user-business-unit']));
-
-            if (!$businessUnit) {
-                throw new InvalidArgumentException('Invalid Business Unit');
-            }
-
-            $user
-                ->setOwner($businessUnit)
-                ->addBusinessUnit($businessUnit);
-        }
-
         if (!empty($options['user-name'])) {
             $user->setUsername($options['user-name']);
         }
@@ -103,6 +111,98 @@ class CreateUserCommand extends ContainerAwareCommand
             $user->setPlainPassword($options['user-password']);
         }
 
+        $this
+            ->setRole($user, $options)
+            ->setBusinessUnit($user, $options)
+            ->setOrganizations($user, $options)
+            ->setProperties($user, $options);
+
+        $this->getUserManager()->updateUser($user);
+    }
+
+    /**
+     * @param User $user
+     * @param array $options
+     * @throws InvalidArgumentException
+     * @return $this
+     */
+    protected function setRole(User $user, $options)
+    {
+        if (!empty($options['user-role'])) {
+            $role = $this->getEntityManager()
+                ->getRepository('OroUserBundle:Role')
+                ->findOneBy(['role' => $options['user-role']]);
+
+            if (!$role) {
+                throw new InvalidArgumentException('Invalid Role');
+            }
+
+            $user->addRole($role);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @param User $user
+     * @param array $options
+     * @throws InvalidArgumentException
+     * @return $this
+     */
+    protected function setBusinessUnit(User $user, $options)
+    {
+        if (!empty($options['user-business-unit'])) {
+            $businessUnit = $this->getEntityManager()
+                ->getRepository('OroOrganizationBundle:BusinessUnit')
+                ->findOneBy(['name' => $options['user-business-unit']]);
+
+            if (!$businessUnit) {
+                throw new InvalidArgumentException('Invalid Business Unit');
+            }
+
+            $user
+                ->setOwner($businessUnit)
+                ->setOrganization($businessUnit->getOrganization())
+                ->addBusinessUnit($businessUnit);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @param User $user
+     * @param array $options
+     * @throws InvalidArgumentException
+     * @return $this
+     */
+    protected function setOrganizations(User $user, $options)
+    {
+        if (!empty($options['user-organizations'])) {
+            foreach ($options['user-organizations'] as $organizationName) {
+                try {
+                    $organization = $this->getEntityManager()
+                        ->getRepository('OroOrganizationBundle:Organization')
+                        ->getOrganizationByName($organizationName);
+                } catch (NoResultException $e) {
+                    throw new InvalidArgumentException('Invalid organization "' . $organizationName .
+                        '" in "--user-organizations" parameter');
+                }
+
+                $user->addOrganization($organization);
+            }
+        }
+
+        return $this;
+    }
+
+    /**
+     * @param User $user
+     * @param array $options
+     * @throws InvalidArgumentException
+     * @return $this
+     */
+    protected function setProperties(User $user, $options)
+    {
         $properties = ['email', 'firstname', 'lastname'];
         foreach ($properties as $property) {
             if (!empty($options['user-' . $property])) {
@@ -110,7 +210,7 @@ class CreateUserCommand extends ContainerAwareCommand
             }
         }
 
-        $this->getUserManager()->updateUser($user);
+        return $this;
     }
 
     /**
