@@ -22,7 +22,7 @@ use Oro\Bundle\UserBundle\Migrations\Schema\v1_10\OroUserBundle as PasswordChang
 use Oro\Bundle\UserBundle\Migrations\Schema\v1_15\UpdateEmailOriginRelation as EmailOrigin;
 use Oro\Bundle\UserBundle\Migrations\Schema\v1_15\RemoveOldSchema;
 use Oro\Bundle\UserBundle\Migrations\Schema\v1_15\SetOwnerForEmail;
-use Oro\Bundle\UserBundle\Migrations\Schema\v1_17\AddRelationToMailbox;
+use Oro\Bundle\UserBundle\Migrations\Schema\v1_16\AddRelationToMailbox;
 
 /**
  * @SuppressWarnings(PHPMD.TooManyMethods)
@@ -83,7 +83,6 @@ class OroUserBundleInstaller implements
         $this->createOroAccessGroupTable($schema);
         $this->createOroUserAccessGroupRoleTable($schema);
         $this->createOroAccessRoleTable($schema);
-        $this->createOroSessionTable($schema);
         $this->createOroUserStatusTable($schema);
 
         /** Foreign keys generation **/
@@ -285,21 +284,6 @@ class OroUserBundleInstaller implements
         $table->addColumn('role', 'string', ['length' => 30, 'precision' => 0]);
         $table->addColumn('label', 'string', ['length' => 30, 'precision' => 0]);
         $table->addUniqueIndex(['role'], 'UNIQ_673F65E757698A6A');
-        $table->setPrimaryKey(['id']);
-    }
-
-    /**
-     * Create oro_session table
-     *
-     * @param Schema $schema
-     */
-    protected function createOroSessionTable(Schema $schema)
-    {
-        $table = $schema->createTable('oro_session');
-        $table->addColumn('id', 'string', ['length' => 255, 'precision' => 0]);
-        $table->addColumn('sess_data', 'text', ['precision' => 0]);
-        $table->addColumn('sess_time', 'integer', ['precision' => 0]);
-        $table->addColumn('sess_lifetime', 'integer', ['nullable' => false]);
         $table->setPrimaryKey(['id']);
     }
 
